cmake: Use TARGET_INCLUDE_DIRECTORIES instead of  INCLUDE_DIRECTORIES
Another one for compliance. If the project is included as

    add_subdirectory(Ring-LRC)

In a client. Then the targets will be `ringclient` or
`ringclient_static`. In the case, the client will inherit those
target include directories, but only if they are declared using
`TARGET_INCLUDE_DIRECTORIES`.

There was already some commented code to do it, so I guess someone
at some point tried and gave up. Can't blame anyone for that, this
is indeed confusing and has contradicting documentation. However,
it is a necessary to create a valid package.

Change-Id: I7669058437319d3298500d8d0b539932d1b11515