useraction: Fix VIEW_CHAT_HISTORY visibility.
Make sure the "show chat" button is enabled when really available.

When this code was written, the chat was only available during a
call. Later, it was extended to keed some logs of existing
conversations.

However, since Alpha1, it is also possible to send offline messages
to Ring accounts. To properly determine the state of the "show chat"
button, this also has to be taken into account. This commits adds
an URI type check to the automata. This requires having optional
CM support in one of the reducing pass, which create some patch
noise, but should not destabilize anything.

This commit also fixes 2 compilation warnings introduced by a
trust request commit earlier this week.

Change-Id: I839ac17330bcfcd98fea687b8bcaf96eb7669957