recordingmodel: Complete the code
It was half written before I left. This was known at the time but
left as-is because the parts that worked were needed ASAP.

This model is used to be able to list and delete individual chat
conversations. It doesn't have many roles for now. This commit only
finish what was already started.

This commit also port it to the newer code convention put forward
by the current maintainers:

 * Use std::vector for private structures (modern c++ first)
 * Use 4 spaces
 * Favor `auto` and `{}` instead of the explicit types.
 * Use shorter line length by using an exit early strategy

This commit also drop the "FILE" media code as it doesn't look
like the daemon will support it anytime soon and commented code
is evil.

Change-Id: I32a7cc37d10c5a8bf430a03189e07ce6072c8a8e