Update patch set 1
Patch Set 1: Code-Review+2

People and developers understand and use Contact Requests, not Trust Request, a concept specific to the Ring daemon and daemon API. Contact requests are not daemon Trust requests, they also include structured contact data (in trust requests it's a binary blob).
It's fine to use "trustrequest" only for things that directly refer to the daemon API. As soon as the payload is unserialized or the request manipulated to be presented to the user, it should be a ContactRequest.

It had to be fixed now because it defines the LRC client API. After merged and used we would not have changed it (or at high cost) and it would have caused endless confusion for developers (for instance by using "trust request" in UI or logs by mistake), and then possibly for users.

Patch-set: 1
Label: Code-Review=+2