Update patch set 3
Patch Set 3:

Well, in the GNOME client (mac too, I think), we set the display name to the same as the account alias, I believe, but we don't have an explicit field for it. So it does sort of get used, at least until the profile vCard is transfered... It could still be useful if we allow disabling the use of profiles... though I'm not sure why you wouldn't want a profile, but a display name...

I'm kind of neutral towards this, since this is something that can easily be changed at any point. If it improves things for KDE, I don't mind merging this, I don't think it will affect any other clients.

Patch-set: 3
Label: Code-Review=0