Update patch set 3
Patch Set 3:

It's up to you. As you said, the impact is rather low and can be later changed quite easily. I prefer to keep aliases separated from DisplayName. The SIP/phone use case is rather important for me and a non negligeable portion of my userbase who migrated from SFLPhone-KDE (actually, that's like 90% of the KDE client current users). In that case, there is often multiple accounts with various network settings/VPNs and the alias is useful to reflect that.

If there becomes a way to disable profiles and use raw accounts again, it would probably be a good idea to properly support that in LRC rather than rely on bugs. That will be a different patch. Keeping the display name visible for that corner case is more akin to "a bug with nice side effect"  than a feature.

Feel free to merge or disagree and reject this patch. This is one of the least important of the patchset I did last week. However I would appreciate if it was merged.

Patch-set: 3
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0