Update patch set 1
Patch Set 1:

> But I do not see a reason why a client would like to run with an unselected account

My point was simply that different clients might use different logic to decide what to do when there is no account selected. If we all agree to have the same behaviour, then this should be in LRC, if not, then we can let each client implement it.

Patch-set: 1