RecentModel: filter on account then regex
This patch tries to simplify the filtering logic.

We first check if there is a selected account, and if not then we
filter everything out. Then we check if there is either no
account set on the CM, or that it matches the selected account.
Finally we see if our filter string is contained in some identifier
of the CM or Person.

Change-Id: Ic00d837bfa224177530a110ac9d09617de5c5d0e
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>