Update patch set 4
Patch Set 4: Code-Review-1

I don't think I agree with this change. Especially when the number is dialed as-is using the dialing call and the state machine. It is fine (and recommended) to choose an account before calling this function (it is the fallback after all), but by removing the last line of defense, you will add more bugs than you fix.

This gets worst if a SIP account is selected as the current account. In that case, calling a ring number using the hash just wont work.

I would prefer this patch to be abandoned, it doesn't fix any real/urgent bugs and introduce new ones.

Patch-set: 4
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-1