use user selected account when CM is not specified
When currentDefaultAccount() was being called with no CM as input
(eg: from ther UserActionModel), then the scheme type was being
set to NONE, and thus Ring accounts were not passing the check in
validAccountForScheme(). This is because we can always detect the
scheme for a Ring URI, so Ring accounts don't support the scheme
of type NONE.

However, in this situation we want the user selected account to
be chosen, since there is no other information available.

Change-Id: Ia2d9b02fb3bb0476f044541d276dba768468feb8