add vcard from contact request
- mapToPersonFromIncomingContactRequest : this function creates a
Person from a vCard stored by the daemon; we make sure to ignore any
ContactMethods which might be contained in the vCard and only use
the CM from which we receive it.

- PendingContactRequestModel::data now can return :
  * PEER_ID : get the username if possible or ring Id if not.
  * FORMATTED_NAME : get the profile name stored in the vCard.
  * COUNT__ : used to get the number of elements in the enum.

- ContactRequest now store a person. This person matches the vCard.

- PeerProfileCollection updates the CM(s) of the vCards it stores
with the last modified date of the vCard in to get the date at which
the contact request was accepted.

[SS: update commit message]
[SS: change how we restore the ContactRequest accept date]

Change-Id: I3bafda5b38d5e2332e095cd5f1f1d0b101847f86
Signed-off-by: Stepan Salenikovich <stepan.salenikovich@savoirfairelinux.com>
Reviewed-by: Stepan Salenikovich <stepan.salenikovich@savoirfairelinux.com>