uri: Allow to set the scheme later.
When the user enters a string, it isn't always yet known if it is a
search for registered name, a SIP URI or some local DNS names.

However at some point the choice can be noarrowed down. This is a good
time to set the scheme. This help later on top pick the right account
and hopefully resolve the RingId before placing the call.

Change-Id: Iad8283fc4c6ff2e4cfe189bf216f00720202ce92
Reviewed-by: Stepan Salenikovich <stepan.salenikovich@savoirfairelinux.com>