call: Prevent ambiguous Ring URI from being used in placeCall()
As of now, unresolved registered names could be passed to placeCall
with a Ring account. As the URI parser doesn't have the metadata
required to correctly pick the scheme, it used `sip:`. This isn't
correct when the account is a Ring account (but the URI is unaware
of that).

This commit set the correct scheme before calling placeCall.

Change-Id: Ied3cfd1ef7f7171e3bf05d51a245e35e2df95ffb
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>