RecentModel: initialize selection in constructor
If there are any ongoing calls during the construction of the
RecentModel we want to make sure to sync the selection with the
CallModel, which will normally select the most recent Call.

This is mostly important for clients which run as seperate processes
from the daemon, and so its possible to have ongoing calls at the
time the client is initialized.

Change-Id: I1edd595ee77c03ca294baf44802f00065e8bca2b