RecentModel: move Calls from CM to Person
Make sure child Calls of a ContactMethod node are moved to the
Person node when a Person node replaces a CM node.

This should prevent two bugs. One is Calls not appearing in the
smart list. The second is the model not properly signaling the
tree view about the child rows moving.

Change-Id: Ibff9186e8494d930726ccb0e8a0e61d2119bc315
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>