call: send vCard only after the call starts
Instead of sending the vCard during an incoming or ringing call,
we wait to send the vCard only once the call has started. This
is to prevent connectivity failures during the call initialization
due to dropped vCard packets.

Note that ::start() is also called after a call goes off hold, so
the vCards will also be sent at this time.

Change-Id: I08e63bf35724ed543c36d88cd43f872aa246b73c