Person: implement Ring::Role::Number role
It returns the bestId() of the lastUsedContactMethod().
This way views will not get an empty QVariant when asking for the
number role from a Person item.

This also makes sense as what we want to show most of the time and
most of the time a Person will only have one ContactMethod anyways.

Change-Id: I0edea8c4fe3f6848a523c4f88d3115cf13b20938
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>