Update patch set 5
Patch Set 5: Code-Review-2

mmm, maybe to discuss with Guillaume, but the logic behind how the banned contacts are fetched is backwards IMHO...

The purpose of the PersonModel is to keep track of all the vCards (well, Person objects, which are generally from vCards). What the BannedPeopleProxy will do, is every time there is a change in the in the PersonModel, it will loop through *every* Person, and then make a call to the daemon to see if it's ContactMethod is banned or not.

The main issue with this (aside from potentially massive amounts of calls to the daemon) is that its not the job of the PersonModel to keep track of all the (daemon) contacts. If for some reason a banned contact is not there, then it will not show up. This could be because a vCard was deleted at some point, or because of a bug in LRC somewhere, or something else.

There are already daemon APIs which give us the list of certificates it has or else the new ones which give us a list of contacts per account.

Actually, in LRC, you should already be able to get the list of banned certificates from Account::bannedCertificateModel. See more in CertifiateModel. Or else we should create a new model which uses the new contact APIs in the daemon.

Afterwards, for display purposes, we can fetch the Person by doing ContactMethod->contact(), but it should not be the other way around...

See this draft for an idea, fyi, I started doing this at some point:
https://gerrit-ring.savoirfairelinux.com/#/c/4413

Patch-set: 5