RecentModel: fix ContactMethod node removal
This fixes one particular instance of "contact duplication" in the
RecentModel. When we had a call from a peer incoming for the very
first time, a ContactMethod ndoe was first being created. Then we
received the profile from the peer, and so a Person node was created.
However, the contactChanged signal could be called before the Person
node is created; in this case the ContactMethod node was not being
removed and the result was a duplication of the the same contact
twice in the RecentModel: once as a CM and once as a Person.

Calling slotLastUsedTimeChanged() on the new Person ensures that
the Person node is created.

We also call slotContactChanged() from slotPersonAdded() to reduce
code duplication.

Change-Id: Ia9b49ab939a017052cccc69f8d72df5db0aef41d