Update patch set 4
Patch Set 4:

> is not clearly indicated (mostly)

Of course. Fixing this should be a priority. I will take care of it ASAP.

> doesn't have sens (a class X is linked to a class Y that have no common stuffs).

Also true. Many of them are no longer needed, but the `friend` line was not removed when the class dependency graph was improved. 

On an historical note, I originally used an LLVM-tidy `RecursiveASTVisitor` based python script to auto-add the d_ptr when I compiled the newly split (public/private) API. This was fast as the compiler auto-updated the changes and patched the code, but that solution is also quite dumb as the fixit plugin is clueless about that really happens. I added way too many firends just to make it compile, only to have the loop fixed a few revisions later. I missed some instances and they ended up being committed.

As for the split itself, there is a long Gerrit comment I wrote for Stepan in 2015, but I can't find it right now. It goes like this (better find the original post, this is a summary):

When possible, the class have their private component in the `.cpp` and they are fully private without any dependencies.

For instances like ::Call<->::CallModel where the coupling is done or purpose (the Factory and Manager patterns) there is a `private/call_p.h` class shared between the 2 entities. A solution could be to merge call.cpp into callmodel.cpp, but that would create a very large and unmaintainable mess of responsabilities.

Another pattern where `friend` is done on purpose is the Placeholder pattern. Much to Edric hatred, this is used in LRC to solve the issue where some network based data sources (like LDAP) take minutes to be loaded while some other (like the history) only takes seconds. As LRC is designed (on purpose) as an asynchronous relational database (see the design document), it has to support indexing partial data. This sucks but was an hard requirement back when the LDAP support was critical. It is still necessary for GMail contacts and other async collections (including the daemon nameservice requests). A placeholder object is a smart pointer that replaces it's data while offering a seamless API with the real object. It is also used by the deduplication algorithm, but this is off topic. I am aware of the discussions to move the relational information to SQLite. However, I have to warn that it wont work. This is exactly what Akonadi did and it eventually killed the project. Keeping relational data in sync with asynchronous sources seems logical, but has been proven to be nearly impossible to implement in practice. As a second example, the big data movement also went in the same direction with their dataset with identical reasoning.

A third place where this is added on purpose, but could be improved, is for the internal state protection. LRC is backed by a series of state machines and exposes "guaranteed" transactions in it's API. The most important of such transaction are the models "beginInsertRows" and "endInsertRows". This is critical to the Gnome client to display the correct values (and not crash on double pointer dereferencing). The KDE client is a little less sensitive to them, but they still cause bugs when broken. There is also many proxies on top of the model to filter or sort them. Those models are very sensitive to the internal state.

Thus, all methods that could cause the state to be affected when called directly have been moved to the private API. Moving them back to the public one carries risk of crashes if the developer was to use them. However, this isn't implemented as cleanly as it could. The private APIs are usually implemented in the same class as the attributes and those class are "all public" as they mix "I need to get secret attributes", "I need to connect to register secret observer object" and "I need some special mutator methods that makes no sense for the public API". In a perfect world, those use case should have different level of internal visibility. However it takes a lot of time and the return on investment is debatable.

A fourth type is the "view" pattern (the database view pattern, not the MVC one, https://en.wikipedia.org/wiki/View_(SQL) ). Some models offer a generic overview of their content, but they are not suitable to the clients. Most of this use case are covered by model proxies, but not all topological disposition are easy to generate using proxies. The PhoneDirectoryModel isn't suitable to generate the BookmarkModel or the RecentModel as displayed by the client. Therefor, some models have "raw" access to one of the main models to do it's own view of their data (and potentially data from multiple models). Some models, like the RecentModel, try to reverse engineer the internal data rather than use it directly, but those strategies have proven to be unsustainable as complexity increases (see the recentmodel commit log).

Finally, a fifth case where `friend` are used are all the bad ones. Some classes have way too many responsibilities and should benefit from a separation of concern based refactoring. TextRecording and ContactMethod are among the worst offenders. The former was rushed weeks before my departure and the later grew organically over the years to become a GodClass as more and more metadata was required to feed the "social" and "analytic" features. I could fix both, but this would break the API quite badly. If you wish to go ahead with both, say so and it will be done. All I ask in return is that the patches be reviewed and merged withing a week as they will be very hard to rebase.


I will perform an analysis of the dependency graph and categorize all `friend`s instances based on the criteria described above. Please put this patch and its sibling on hold until I am done. I should deliver this review soon. We will have more information to take enlighten engineering decision instead of the current piecemeal puzzle.

Patch-set: 4