Update patch set 1
Patch Set 1:

Just a word of warning. The reason it's done that way is that ContactMethods should only be created by their factory (PhoneDirectoryModel::getNumber()). If stray contactmethods are created and used as regular ones, the indexes get corrupted and the models give garbage.

I know some patches recently created some of them, but I that regressed quite many models, added duplicates in the RecentModel, including crashy ones, broke the deduplication algorithm as the duplicates are not indexed and broke the phone numbers stored in the vcards because the factory generated a copy because it wasn't aware of the original object. Stepan cleaned up some of it, but I think there is a few remaining issues.

In this case it there wont be much consequences beside a few crashes on older code, but something needs to be done to avoid the black methods being filled with data and inserted into the indices. as you say they are mutable and there is like C++98 could do about it. Sure, the BLANK was a quick hack, even if it was inserted into an index, at least it would pollute it only once. Now there is the possibility that many of them will end up being added.

Also, on a related note, what do you think about splitting that class in 2. I would like to move all the statistic properties into their own object. That would help with to reduce the coupling of this class. ContactMethod is getting way too large.

Patch-set: 1
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0