fix LIB_EXPORT behavior and misuse
This patch fixes DLL symbols export and especially imports from LRC.
When included in Windows client, classes were not using Q_DECL_IMPORT
which, on this platform, specify that symbols are imported from a lib.

A consequence of that is that Qt connects with the new syntax were
failing thus forcing us to rely on old-style ones. Those ones are not
checked at compile time and not compatibles with lambdas.

It may also fix other issues related to symbols availability and
uniqueness.

Change-Id: I2578deca5215a7e17327405ae6462c373b229ada