ensure daemon is running before Lrc construction
On non-DBus platforms, where the daemon is not running in its own
process, LRC is responsible for its initalization. This patch ensures
it is done in Lrc class constructor before any attempt to grab data
from it (like accounts).

It then makes it possible to use the new models in Win32 and macOS
clients.

Change-Id: Ib587bf86e17ae7f19766ac5bb5d8c91aa05d335c