edit data namespaces
- after working on the WIP branch we need to get these edits before
the ones for the models. This patch edits the following namespaces :
account, contact, conversation and messages.

- add functions to convert type and status to string and vice versa.

- edit Info structures, remove unused members and add some needed.

- add profile namespace. Informations stored by daemon or database
about account or contact will be accessible trough profile. Other
informations will be considered as volatil.

- message namespace are renamed interaction namespace. The reason,
is a message is only a part of the different interaction that could
be present in the historic.

Change-Id: Iba46194eb29ff394481663bef4ca07dc4bd899f6