Update patch set 25
Patch Set 25:

> hmm... I'm not sure its a good idea to force classes to have a
 > behavioral instance in their ctor.
 > There is too much coupling...
 > Alternative?

These signals are for all accounts and it makes easy the dev of the client.
The alternative can be to have signals directly in the ConversationModel like before but it's not the role of the ConversationModel to emit signals like showChatView and showChatView can come from another account.

Patch-set: 25
Label: Code-Review=0