newcallmodel: remove unneeded argument
A boolean argument in toggleMedia() was passed specifying if the
chosen media was to be muted or not. This is not consistent with the
other toggle*() methods used to manage calls.

Moreover, it is not needed at all as we can retrieve the current state
of a media and toggle it from LRC itself, thus removing the need to do
it on each client.

Change-Id: I53528d925c8336b92afc47367adef971d11e9a25