add height and width in Video::Frame structure
By adding those member variables, this information is now conveyed from
daemon to client on a per-frame basis. This is necessary as ffmpeg may
output frames with a size different than what is returned by
Renderer::size().

This later value is set when a video stream reception is established
with height and width corresponding to format of frame sent by the
other end. After decoding, this first size may not be equal to the size
of the frame decoded by ffmpeg (seen at least on macOS with hardware
acceleration enabled).

Finally, doing so is more secure as we are protected by the Renderer
mutex. Previously, getting size and frame were 2 different operations
(method calls) with no guarantee of atomicity, and thus consistency.

Change-Id: I05699b71d2afd38deeea36374830ff00d7d66ad7