make addContact() preserve profile::Type
On non-DBus platform, when trusting a pending contact. The call from
daemon to ContactModelPimpl::addToContacts happens before the end of
ContactModel::addContact. This has the effect of switching the profile
type to RING then back to PENDING.

We now ensure that profileInfo.type is set according to the value
already present in pimpl_->contacts if any.

Change-Id: I4609056fe9191d00affe93ef1223bd06073b2276