temporary objects redux
the temporary objects (conversation and contact) were
totally functionnal only on the GTK client through a hack.

In order to make the temporary contact and conversation
functional outside the GTK client, we had to change its
semantics all the way through its processing.

The major changes are:
  - the temporary conversation has now a UID set when a
	name is found on the nameservice, allowing the
	conversation to be interactive,
  - sending a message, or calling said conversation correctly adds
	the new contact (no duplication or ghost element).

/!\
these changes have to be ported to the clients to function
/!\

Change-Id: I7f010a6dc20aeab9099864350a20dee459c1449c