fix: sort and deduplicate frame rates
In the VideoDevicePrivate constructor, rates are appended to the
res->d_ptr->m_lValidRates list of each resolution in a foreach
loop, without removing duplicates or sorting.

In this commit we add a check to the foreach loop which avoids
adding an element already present in the list. In addition to
that, once all rates have been added to the list, we sort the
resulting res->d_ptr->m_lValidRates in increasing order using
qsort.

Change-Id: I4e696d03119b03b37cddef9005464fdca78818c5
Tuleap: #1850
Reviewed-by: Sébastien Blin <sebastien.blin@savoirfairelinux.com>