fix: race cond. when deleting accounts, lrc side
Currently, whenever an account is removed the lrc first frees the
structures associated with the account with erase() and then notifies
the client (using accountRemoved signal). This is a major issue
because in the time span between the erase() call and the execution of
accountRemoved callback by the client, *all* client structures
pointing to the LRC-side structures of the removed account are left
dangling, leading to various potentially security relevant NULL
pointer dereference and use-after-free issues.

In this patch we:
 * first signal, then wait for the client to stop using the removed
   account's structures and then call erase()
 * introduce an *API BREAKING* synchronization mechanism (1) in order
   to allow the LRC to wait for the client to stop using the account
   structures before freeing them.

(1): After updating its account pointers and doing relevant changes,
the client MUST now call the flagFreeable(accountId) function. This
allows the LRC to be notified that the client has stopped using the
removed account's structures and that it can now free them safely.

For now, the API breaking changes are deactivated by default on all
platforms apart from Linux (see CHK_FREEABLE_BEFORE_ERASE_ACCOUNT
option), but this should only be considered as a temporary transition
measure.

Change-Id: I9c5422524d09afcdf471e288edf462df8c382ad7
Reviewed-by: Sébastien Blin <sebastien.blin@savoirfairelinux.com>