Do not add a banned account to contacts
Currently it is possible to add a banned account to contacts. This
is a misfeature and should not be allowed. Whenever addContact is
called with a banned contact, log and return.

Also: In order to avoid crashes, we should really check the contact
index variable before using it in conversationmodel. In this patch
we add checks to sendMessage and placeCall so we stop trying to
contact an account which is not in the contact list.

Change-Id: I038c7049781a9bce951f09ed4a0f1d3100e6df3c