datatransfer: fix deadlocks for file from pending contact
Since 35d142b354223662570fb7dc7f6b6e8081deae26, a deadlock appeared
when an unknown contact send a file. So, to fix it, we need to avoid
to call signals inside a lock and to pass the created signal through
contactmodel first to create the new pending contact like calls and
text messages.

Change-Id: I3a92d553e8653477f85e7e12ce92796f397caa52
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>