newdevicemodel: deprecate ringdevicemodel in favor of newdevicemodel
Ring devices are currently managed via a Qt model. With this patch,
this class is now deprecated in favor of NewDeviceModel linked to
the AccountModel used by clients.

This patch add the ability to revoke devices and rename current used
device for LRC based clients.

Add some unit-tests in test/newdevicemodeltester.*

Change-Id: Ifcec4e573f92967d7e0ee8759f1159328957c7ba
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>