contactmodel: introduce bannedStatusChanged signal
cb08d76 introduced a new banned contacts logic to the contact model,
which allows to manage banned contacts without using the old lrc.
While this patch is fully functional, the problem of how exactly the
client is notified about a banning / unbanning operation is still
unclear: currently the only way the client gets notified is via the
refreshFilter() operation. This is enough for managing banned
contacts in the smart list, but not in the banned contacts view.

In order to addres this problem we introduce a new signal,
bannedStatusChanged, which allows the client to be notified when a
contact is banned or unbanned.

Change-Id: I2ea845e0fe311048479777504b74885cb931aa13
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>