conversationmodel: fix race condition
Fix race condition when sending message to temporary contacts.

In order to address this issue we introduce a synchronization
mechanism that waits for slotContactAdded to be called before
processing further.

TODO: This patch fixes the issue for sendMessage only, not for
placeCall & Co.

KNOWN ISSUES: The wait call freezes the application, that is if
slotContactAdded never comes or takes a lot of time to execute, the
application might freeze.

Change-Id: Ie2248d4f1bfb7a7c495c2dc63441d2255bd5e7e0