conversationmodel: concurrency issue in placeCall
If the contactAdded signal has a small delay and comes after the
if (isTemporary && (contactIndex = indexOfContact(convId)) < 0)
check from line 538, the call is abruptly stopped without valid
reasons.

In this patch we introduce the same synchronization mechanisms as
af04c34 so that we can handle small delays of contactAdded nicely
and still stop call properly in case of failure while adding contact.

Change-Id: I83b7cd5910d21841b09e969d3baa7b484906941b
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>