conversationmodel: send file to create contact
The sendFile() method contains contradictory code: "if
(peerUri.empty()) return" forbids sending files to non contacts, but
a line later "pimpl_->sendContactRequest(peerUri)" is called even if
this line can only be executed in the case where peerUri is already a
contact.

In this commit we allow sending files to non contacts (this makes
sense if peer has already accepted the sender as contact), and
protect it with the same synchronization mechanism as af04c34.

Change-Id: I25432fcfdd3f46068add27009a77108ac75d3124
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>