newaccountmodel: stop using old lrc for accountsChanged signal
Add a CallbackHandler entry for deamon signal accountsChanged.

Add an updateAccounts() function to the NewAccountModel and connect it
to this new signal. This function detects new accounts and account
removals. It calls addToAccounts (synchronous) and slotAccountRemoved
(asynchronous) when appropriate.

Stop using registrationStateChanged signal to detect new accounts.

Add a new signal internal to the NewAccountModel which is used to
trigger slotAccountRemoved asynchronously.

Change-Id: Id774df852e63b69658160473962ec499fc244c53