Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> I think this is overkill if you only want to update the register username.
> 
> By the way, not sure what bug do you want to see resolved with this patch?

It' s not a bug, but a registered name not is visible right away, after registration finished. This patch help to update account with a new name, but it is not fix the case when application was closed.

Patch-set: 1
Reviewer: Gerrit User 225 <225@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0