ringing timeout: implement let a message feature
Use dedicated states to represent BUSY and PEER_BUSY daemon status.

Do not emit showChatView when call ends in PEER_BUSY status. In this
case emit newly added showLeaveMessage signal. When receiving this
signal, the client should ask the user whether it wants to record a
message.

Continue to emit showChatView in the BUSY state.

Change-Id: I03882ada77366332443b835760034c8e370d9694
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>