conversationmodel: store previous call id
At the end of a call slotCallEnded is executed and sets the
conversation's callId to "". In some cases however, the previous
callId might be needed by the conversationmodel, and as such it should
be stored in the conversation info struct.

Also, do not call selectConversation() in slotCallEnded and cleanup
slotCallStateChanged() in newcallmodel.

Change-Id: I9bb0de0ff6ec1378e6ca59470fc1a5ee186cf0ba
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>