call management: avoid signal duplication
At the end of a call the state machine goes through several state
changes. The daemon guarantees that we will first go through HUNGUP,
INACTIVE, BUSY or FAILURE, and then through OVER.

Currently FAILURE is considered to be the same as OVER by the LRC.
This is wrong because FAILURE is a transitional state just like
HUNGUP, not a final state like OVER. Because of that FAILURE should
translate to TERMINATING and not ENDED.

Also, we should always emit a signal during the first state
transition, that is when we switch to HUNGUP, INACTIVE, BUSY or
FAILURE, not when we switch to OVER. It is not meaningful to emit
signals in both cases because it might result in contradictory
actions, graphical glitches and obviously in a waste of resources.

Change-Id: I2356ef133ba504555314d603a80bcebf3d112e02
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>