call state machine: fix SEARCHING state transition
The call state machine expects a two step state transition: the
daemon should first transition to a terminating state, and then to
OVER. This is almost always the case, expect for the SEARCHING state.

In fact this state is a bit special since it does not correspond to
any well defined daemon state. In order to avoid issues resulting
from an invalid SEARCHING -> ENDED transition, add a temporary state
transition to TERMINATING.

Change-Id: I074d9a1beba2e39ef3089733485b57311e834214
Gitlab: #392