conversationmodel: fix deadlock on first interaction
af04c34 introduced a condition variable used to block until the
conversation was created, circumventing a race condition when
attempting to insert interactions. However, when not using D-Bus,
slotContactAdded will be invoked on the blocked thread, and thus
deadlock until the wait_for timeout.

This patch replaces the synchronization mechanism with callbacks.

Change-Id: I26141a31853b526391a01b176d99135d8738be18