conversationmodel: fix deadlock/race condition on first interaction
af04c34 introduced a condition variable used to block until the
conversation was created, circumventing a race condition when
attempting to insert interactions. However, if slotContactAdded
is invoked on the blocked thread, the thread will block until
the wait_for's timeout, and then still be susceptible to the initial
race condition.

This patch replaces the synchronization mechanism with callbacks.

Change-Id: I5f975a6245232e586be7df3f29008c1162d6b208