Update patch set 2
Patch Set 2:

Dangerous case:

If a user has for example 2 accounts (say A and B). If a user got a current conversation between account A and B on this device, maybe this conversation will be lost? I don't know the whole implications, but the migration can be more tricky :)

Patch-set: 2
Label: Verified=0