Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> setCredentials already called in NewAccountModel btw

Yeah, but it was called in setAccountConfig, which we do not want to touch for the new sip credentials feature. For this new feature, we need to have a way to directly interact with the setCredential function instead of setting sip account properties.

Also, setAccountConfig function only care about the sip account root crendials instead of its sub-crendials

Patch-set: 2