Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (4 comments)

Thank you for the feedback. Your were right, there was a lot of duplicate code. I consolidated most of it into the same function.
My vision was that eventually we would remove the old function with the struct return value. But it's still a good idea to keep the code to a minimum even for now.

Patch-set: 2