conversations: use collection proxy for filtered and sorted views
Provides a STL container proxy class with methods for generating
sorted and filtered views of the underlying data similar to
qsortfilterproxymodel.

This helps to prevent excessive copying and reduces the complexity
of the code used to manage conversation list filtering/sorting.
The container mutations occur on references and the copy-ctor +
copy assignment operators are now deleted for the conversation
Info type.

Change-Id: I50b01e3c427563d8e32402f97dec79c8dcc865cf