directrenderer: fix multiple possible video crashes
Multiple direct renderers can be created with the same id, each
on their own thread. We probably shouldn't delete the the
underlying frame buffer while it's being accessed by jami daemon
or another DirectRenderer wrapper.

Note: the Shm rendering engine may suffer from a similar bug but
is not fixed in this commit.

Change-Id: I79b23e3b3532a61c282d4ad1a96b6df38cb37130