Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 5:
> 
> nice work so far 😊  2 comments:
> 
> - the issue number seems to be wrong.  did you mean #470 perhaps?
> 
> - this must be accompanied with an option under the 'General' settings view -- similar to the one for displaying images and videos in the chatview -- to allow toggling the feature.  my personal preference would be for it to be disabled by default, as a privacy measure, to avoid contacting random remote servers on user's computer when opening messages.

This was assigned issue #407 in the client-qt project, but this is done in ring-lrc so the number is wrong. I will remove it for now.

For the toggle, I talked to Adrien and he said I could just use the same toggle as the one used to display links/images/videos in the chat view. If it is better to have a separate button, I have no problem implementing that.

Thanks

Patch-set: 6