conversation view: fix adding contact not showing registered name
After a conversation is made permanent, and a contact is added, the
conversation view will reload with the info hash in the header and
message send bar because the conversation becomes ready prior to the
contact being added to the list allowing for profile info queries to
be successful. This patch adds a PENDING contact item that will be
upgraded once the daemon adds the contact.

Gitlab: #479
Change-Id: Ie680bb61803812165f8f061e1cb8c21f77b381f7